+++ scripts/add_electricity.py
@@ -1213,27 +1213,15 @@ if __name__ == "__main__":
     )
 
     tyndp_renewable_carriers = (
-        [
-            subcarrier
-            for carrier in params.electricity["pecd_renewable_profiles"][
-                "technologies"
-            ].values()
-            for subcarrier in carrier
-        ]
+        list(chain(*params.electricity["pecd_renewable_profiles"]["technologies"].values()))
         if params.electricity["pecd_renewable_profiles"]["enable"]
         else []
     )
     if len(tyndp_renewable_carriers) > 0:
         logger.info(
-            f"Skipping renewable carriers '{', '.join(tyndp_renewable_carriers)}'. They will be attached later on with TYNDP data."
+            f"Skipping renewable carriers - they will be attached later with TYNDP data: {', '.join(tyndp_renewable_carriers)}"
         )
-    renewable_carriers = set(
-        [
-            carrier
-            for carrier in params.electricity["renewable_carriers"]
-            if carrier not in tyndp_renewable_carriers
-        ]
-    )
+    renewable_carriers = set(params.electricity["renewable_carriers"]).difference(tyndp_renewable_carriers)
     extendable_carriers = params.electricity["extendable_carriers"]
     conventional_carriers = params.electricity["conventional_carriers"]
     conventional_inputs = {