+++ scripts/clean_pecd_data.py
@@ -37,7 +37,7 @@ logger = logging.getLogger(__name__)
 
 
 def read_pecd_file(
-    node: list,
+    node: str,
     fn_pecd: str,
     cyear: str,
     pyear: str,
@@ -140,8 +140,7 @@ if __name__ == "__main__":
 
     pecd_df = (
         pd.concat(demand, axis=1)
-        .reindex(nodes, axis=1)
-        .fillna(0.0)  # include missing node data with empty columns
+        .reindex(nodes, axis=1, fill_value=0.0)  # include missing node data with empty columns
         .rename(
             columns=lambda x: x.replace("UK", "GB")
         )  # replace UK with GB for naming convention